updated custom int type

+++ src/Percentage.cpp
@@ -2,8 +2,8 @@
 
 namespace sgrid {
 
-Percentage operator""_pct(u64 value) {
-  return value;
+Percentage operator""_pct(unsigned long long value) {
+  return static_cast<u64>(value);
 }
 
 } // namespace sgrid
